refactor: group bn256 ext-field into ext_field
